Refactor #3 - change icon & image component's file path
