+++ project/dynamic-busan/src/js/component/Icon.js
@@ -1,4 +1,4 @@
-import { createElementNS } from './util/dom';
+import { createElementNS } from '../util/dom';
 
 /**
  * @description Drop Down SVG Icon을 생성합니다.
